[feature/#1011] snackBar 호출 위치 수정
