+++ app/src/main/java/org/sopt/official/feature/auth/AuthScreen.kt
@@ -5,12 +5,20 @@ import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.padding
 import androidx.compose.material3.Scaffold
+import androidx.compose.material3.SnackbarHost
+import androidx.compose.material3.SnackbarHostState
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.rememberCoroutineScope
+import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.unit.dp
 import androidx.navigation.NavHostController
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.rememberNavController
+import kotlinx.coroutines.launch
 import org.sopt.official.designsystem.SoptTheme
+import org.sopt.official.feature.auth.component.CertificationSnackBar
 import org.sopt.official.feature.auth.feature.authmain.navigation.AuthMainNavigation
 import org.sopt.official.feature.auth.feature.authmain.navigation.authMainNavGraph
 import org.sopt.official.feature.auth.feature.certificate.navigation.certificationNavGraph
@@ -20,10 +28,34 @@ import org.sopt.official.feature.auth.feature.certificate.navigation.navigateCer
 internal fun AuthScreen(
     navigateToUnAuthenticatedHome: () -> Unit,
     onGoogleLoginCLick: () -> Unit,
-    navigateToChannel:() -> Unit,
+    navigateToChannel: () -> Unit,
     navController: NavHostController = rememberNavController(),
 ) {
+    val snackBarHostState = remember { SnackbarHostState() }
+    val coroutineScope = rememberCoroutineScope()
+    val onShowSnackBar: () -> Unit = remember {
+        {
+            coroutineScope.launch {
+                snackBarHostState.showSnackbar("인증번호가 전송되었어요.")
+            }
+        }
+    }
+
     Scaffold(
+        snackbarHost = {
+            Box(modifier = Modifier.fillMaxSize()) {
+                SnackbarHost(
+                    hostState = snackBarHostState,
+                    modifier = Modifier
+                        .padding(top = 16.dp)
+                        .padding(horizontal = 16.dp)
+                        .align(alignment = Alignment.TopCenter),
+                    snackbar = { message ->
+                        CertificationSnackBar(message = message.visuals.message)
+                    }
+                )
+            }
+        },
         modifier = Modifier.fillMaxSize(),
         content = { paddingValue ->
             Box(
@@ -47,9 +79,10 @@ internal fun AuthScreen(
                             navController.navigateCertification()
                         },
                         navigateToChannel = navigateToChannel
-
                     )
-                    certificationNavGraph()
+                    certificationNavGraph(
+                        onShowSnackBar = onShowSnackBar
+                    )
                 }
             }
         }
