+++ app/src/main/java/org/sopt/official/feature/auth/feature/certificate/CertificationScreen.kt
@@ -36,26 +36,22 @@ import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.material3.SnackbarHost
-import androidx.compose.material3.SnackbarHostState
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.hilt.navigation.compose.hiltViewModel
-import androidx.lifecycle.Lifecycle
-import androidx.lifecycle.LifecycleOwner
 import androidx.lifecycle.compose.LocalLifecycleOwner
 import androidx.lifecycle.flowWithLifecycle
-import kotlinx.coroutines.launch
+import org.sopt.official.R
 import org.sopt.official.R.drawable.ic_auth_memeber_error
 import org.sopt.official.R.drawable.ic_auth_process
+import org.sopt.official.common.view.toast
 import org.sopt.official.designsystem.Blue500
 import org.sopt.official.designsystem.BlueAlpha100
 import org.sopt.official.designsystem.Gray10
@@ -68,45 +64,26 @@ import org.sopt.official.designsystem.White
 import org.sopt.official.feature.auth.component.AuthButton
 import org.sopt.official.feature.auth.component.AuthTextField
 import org.sopt.official.feature.auth.component.AuthTextWithArrow
-import org.sopt.official.feature.auth.component.CertificationSnackBar
 import org.sopt.official.feature.auth.component.PhoneCertification
 
 @Composable
 internal fun CertificationRoute(
+    onShowSnackBar: () -> Unit,
     viewModel: CertificationViewModel = hiltViewModel()
 ) {
-    val snackBarHostState = remember { SnackbarHostState() }
-    val coroutineScope = rememberCoroutineScope()
-    val onShowSnackBar: () -> Unit = remember {
-        {
-            coroutineScope.launch {
-                snackBarHostState.showSnackbar("인증번호가 전송되었어요.")
-            }
-        }
-    }
-
     val lifecycleOwner = LocalLifecycleOwner.current
+    val context = LocalContext.current
 
     LaunchedEffect(viewModel.sideEffect, lifecycleOwner) {
         viewModel.sideEffect.flowWithLifecycle(lifecycle = lifecycleOwner.lifecycle)
             .collect { sideEffect ->
                 when (sideEffect) {
                     is CertificationSideEffect.ShowToast -> {
-                        snackBarHostState.showSnackbar(sideEffect.message)
+                        context.toast(sideEffect.message)
                     }
                 }
             }
     }
-
-    SnackbarHost(
-        hostState = snackBarHostState,
-        modifier = Modifier
-            .padding(top = 16.dp)
-            .padding(horizontal = 16.dp),
-        snackbar = { message ->
-            CertificationSnackBar(message = message.visuals.message)
-        }
-    )
 
     CertificationScreen(
         onCreateCodeClick = {
@@ -122,74 +99,82 @@ internal fun CertificationRoute(
 @Composable
 private fun CertificationScreen(
     onCreateCodeClick: () -> Unit,
-    onCertificateClick:()-> Unit
+    onCertificateClick: () -> Unit
 ) {
-    Column(
-        modifier = Modifier
-            .fillMaxSize()
-            .padding(horizontal = 24.dp),
-        horizontalAlignment = Alignment.CenterHorizontally
-    ) {
-        Spacer(modifier = Modifier.height(72.dp))
-        TopBar()
-        Spacer(modifier = Modifier.height(44.dp))
-        PhoneCertification(
-            onPhoneNumberClick = onCreateCodeClick,
-            textColor = Gray80
-        )
-        Spacer(modifier = Modifier.height(10.dp))
-        AuthTextField(
-            modifier = Modifier.fillMaxWidth(),
-            labelText = "",
-            hintText = "인증번호를 입력해 주세요.",
-            onTextChange = {},
-        )
-        Spacer(modifier = Modifier.height(41.dp))
-        AuthButton(
+    Column {
+        Image(
+            painter = painterResource(id = R.drawable.ic_auth_arrow_left),
+            contentDescription = "뒤로가기 버튼",
             modifier = Modifier
-                .fillMaxWidth()
-                .border(
-                    color = Blue500,
-                    width = 1.dp,
-                    shape = RoundedCornerShape(10.dp)
-                ),
-            padding = PaddingValues(vertical = 14.dp, horizontal = 18.dp),
-            onClick = {},
-            containerColor = BlueAlpha100,
+                .padding(vertical = 12.dp)
+                .padding(start = 20.dp)
+        )
+        Column(
+            modifier = Modifier
+                .fillMaxSize()
+                .padding(horizontal = 24.dp),
+            horizontalAlignment = Alignment.CenterHorizontally
         ) {
-            Row(horizontalArrangement = Arrangement.spacedBy(10.dp)) {
-                Image(
-                    painterResource(id = ic_auth_memeber_error),
-                    contentDescription = "에러 아이콘",
-                )
-                Column {
-                    AuthTextWithArrow(
-                        text = "SOPT 회원 인증에 실패하셨나요?",
-                        textStyle = SoptTheme.typography.body14M
+            TopBar()
+            Spacer(modifier = Modifier.height(44.dp))
+            PhoneCertification(
+                onPhoneNumberClick = onCreateCodeClick,
+                textColor = Gray80
+            )
+            Spacer(modifier = Modifier.height(10.dp))
+            AuthTextField(
+                modifier = Modifier.fillMaxWidth(),
+                labelText = "",
+                hintText = "인증번호를 입력해 주세요.",
+                onTextChange = {},
+            )
+            Spacer(modifier = Modifier.height(41.dp))
+            AuthButton(
+                modifier = Modifier
+                    .fillMaxWidth()
+                    .border(
+                        color = Blue500,
+                        width = 1.dp,
+                        shape = RoundedCornerShape(10.dp)
+                    ),
+                padding = PaddingValues(vertical = 14.dp, horizontal = 18.dp),
+                onClick = {},
+                containerColor = BlueAlpha100,
+            ) {
+                Row(horizontalArrangement = Arrangement.spacedBy(10.dp)) {
+                    Image(
+                        painterResource(id = ic_auth_memeber_error),
+                        contentDescription = "에러 아이콘",
                     )
-                    Spacer(modifier = Modifier.height(10.dp))
-                    Text(
-                        text = "번호가 바뀌었거나, 인증이 어려우신 경우 추가 정보 인증을 통해 가입을 도와드리고 있어요!",
-                        color = Gray60
-                    )
+                    Column {
+                        AuthTextWithArrow(
+                            text = "SOPT 회원 인증에 실패하셨나요?",
+                            textStyle = SoptTheme.typography.body14M
+                        )
+                        Spacer(modifier = Modifier.height(10.dp))
+                        Text(
+                            text = "번호가 바뀌었거나, 인증이 어려우신 경우 추가 정보 인증을 통해 가입을 도와드리고 있어요!",
+                            color = Gray60
+                        )
+                    }
                 }
             }
+            Spacer(modifier = Modifier.weight(1f))
+            AuthButton(
+                modifier = Modifier.fillMaxWidth(),
+                padding = PaddingValues(vertical = 16.dp),
+                onClick = onCertificateClick,
+                containerColor = Gray10,
+                contentColor = Gray950,
+                disabledContentColor = Gray60,
+            ) {
+                Text(
+                    text = "SOPT 회원 인증 완료",
+                    style = SoptTheme.typography.body14M
+                )
+            }
+            Spacer(modifier = Modifier.height(32.dp))
         }
-        Spacer(modifier = Modifier.weight(1f))
-        AuthButton(
-            modifier = Modifier.fillMaxWidth(),
-            padding = PaddingValues(vertical = 16.dp),
-            onClick = onCertificateClick,
-            containerColor = Gray10,
-            contentColor = Gray950,
-            disabledContentColor = Gray60,
-        ) {
-            Text(
-                text = "SOPT 회원 인증 완료",
-                style = SoptTheme.typography.body14M
-            )
-        }
-        Spacer(modifier = Modifier.height(32.dp))
     }
 }
