+++ app/src/main/java/org/sopt/official/feature/auth/feature/certificate/navigation/CertificationNavigation.kt
@@ -17,9 +17,12 @@ fun NavController.navigateCertification(
 }
 
 fun NavGraphBuilder.certificationNavGraph(
+    onShowSnackBar: () -> Unit
 ) {
     composable<CertificationNavigation> {
-        CertificationRoute()
+        CertificationRoute(
+            onShowSnackBar = onShowSnackBar
+        )
     }
 }
